Refactor: Migrate JsonUtil to Java and update dependencies
- Migrated `JsonUtil.kt` to `JsonUtil.java` for compatibility.
- Updated Gradle wrapper to version 8.14.3.
- Updated Android Gradle Plugin to 8.13.0.
- Updated compileSdk and targetSdk to 36.
- Added `im-core` module as a local AAR dependency.
- Made minor adjustments to SDK calls for type compatibility.
- Removed debug-only logging in `LogcatHelper`.
- Commented out `deleteMessageList` and `setGlobalRecvMessageOpt` in `MessageManager` (marked with TODO).

+++ OpenIM-SDK/src/main/java/io/openim/android/sdk/manager/ConversationManager.kt
@@ -84,7 +84,7 @@ class ConversationManager {
         return Open_im_sdk.getConversationIDBySessionType(
             ParamsUtil.buildOperationID(),
             sourceId,
-            sessionType
+            sessionType.toLong()
         )
     }
